Test the env var sourcing -> decimals test should pass

+++ test/src/test/java/org/hiero/mirror/test/e2e/acceptance/client/ContractClient.java
@@ -36,6 +36,11 @@ public class ContractClient extends AbstractNetworkClient {
 
     @Override
     public void clean() {
+        if (Boolean.parseBoolean(System.getenv("CI"))) {
+            // In CI we don't want to cleanup as the entities are needed in the k6 test in the next step.
+            log.warn("Acceptance tests running in CI -> skip cleanup.");
+            return;
+        }
         log.info("Deleting {} contracts", contractIds.size());
         deleteAll(contractIds, id -> deleteContract(id, client.getOperatorAccountId(), null));
     }
